# GoMailFlows/docs/openapi.yaml
openapi: 3.1.0
info:
  title: HiTechCloud MailFlows API
  version: 1.0.0
  description: |
    HiTechCloud MailFlows - Enterprise Email SaaS Platform API
    
    ## Features
    - **Domain Management**: Register and manage email domains with automatic DNS provisioning
    - **Email Accounts**: Create and manage email mailboxes
    - **SMTP Configuration**: Per-domain or global SMTP relay configuration
    - **JMAP Support**: Full JMAP protocol (RFC 8620/8621) for modern email access
    - **Email Marketing**: Campaign management, subscriber lists, analytics
    - **SaaS Packages**: Flexible service plans with granular feature control
    - **Admin Panel**: Complete system administration capabilities
    - **DNS Templates**: Customizable DNS record templates for domain provisioning
    - **MCP Integration**: Model Context Protocol support for AI-powered management
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.
    Obtain tokens via `/api/v1/auth/login` or `/api/v1/auth/register`.
    
    ## Rate Limiting
    API endpoints are rate-limited. Limits vary by endpoint and user package.
  contact:
    name: HiTechCloud Support
    url: https://hitechcloud.io/support
    email: support@hitechcloud.io
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.mailflows.hitechcloud.io/api/v1
...
security:
  - BearerAuth: []
paths:
...
  /email-accounts:
    get:
      tags: [email-accounts]
      summary: List email accounts
      parameters:
        - name: domain_id
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Account list

    post:
      tags: [email-accounts]
      summary: Create email account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmailAccountRequest'
      responses:
        '201':
          description: Account created
  /email-accounts/{id}:
    put:
      tags: [email-accounts]
      summary: Update email account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmailAccountRequest'
      responses:
        '200':
          description: Account updated

    delete:
      tags: [email-accounts]
      summary: Delete email account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account deleted

  # ============================================================
  # Mail Operations
  # ============================================================
  /mail/folders:
    get:
      tags: [mail]
      summary: List mail folders
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Folder list
  /mail/messages:
    get:
      tags: [mail]
      summary: List messages
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: folder
          in: query
          schema:
            type: string
            default: INBOX
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Message list
  /mail/send:
    post:
      tags: [mail]
      summary: Send email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendEmailRequest'
      responses:
        '200':
          description: Email sent

  # ============================================================
  # Packages
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/register
  schemas:
...
    CreateEmailAccountRequest:
      type: object
      required: [domain_id, local_part, password]
      properties:
        domain_id:
          type: string
          format: uuid
        local_part:
          type: string
          description: Part before @ in email address
        password:
          type: string
          minLength: 8
        display_name:
          type: string
        storage_quota:
          type: integer
          format: int64
          description: Storage quota in bytes
    UpdateEmailAccountRequest:
      type: object
      properties:
        display_name:
          type: string
        is_enabled:
          type: boolean
        storage_quota:
          type: integer
          format: int64
        max_send_per_day:
          type: integer
        forward_to:
          type: string
        auto_reply:
          type: boolean
        auto_reply_text:
          type: string
        password:
          type: string
          minLength: 8
    SendEmailRequest:
      type: object
      required: [account_id, to, subject]
      properties:
        account_id:
          type: string
          format: uuid
        to:
          type: array
          items:
            type: string
            format: email
          minItems: 1
        cc:
          type: array
          items:
            type: string
            format: email
        bcc:
          type: array
          items:
            type: string
            format: email
        subject:
          type: string
        body:
          type: string
        html_body:
          type: string
        reply_to:
          type: string
          format: email
        in_reply_to:
          type: string
        attachments:
          type: array
          items:
            type: object
            properties:
              filename:
                type: string
              content_type:
                type: string
              content:
                type: string
                description: Base64 encoded content
...